+++ Runtime/ResourceManager/Util/MonoBehaviourCallbackHooks.cs
@@ -1,16 +1,12 @@
 ﻿using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.ResourceManagement.Util;
 
-internal class MonoBehaviourCallbackHooks : MonoBehaviour
+internal class MonoBehaviourCallbackHooks : ComponentSingleton<MonoBehaviourCallbackHooks>
 {
     public event Action<float> OnUpdateDelegate;
 
-    void Awake()
-    {
-        DontDestroyOnLoad(gameObject);
-    }
+    protected override string GetGameObjectName() => "ResourceManagerCallbacks";
 
     // Update is called once per frame
     void Update()